Schema switcher for folder metadata
